mini_projects/gabor.torma/backend/requirements.txt
langgraph
langchain
langchain_core
langchain_openai
pydantic
python-dotenv
fastapi
uvicorn
python-multipart
python-docx
requests
chromadb
langchain-chroma
langchain-community
